+++ server.js
@@ -21,8 +21,7 @@ app.use(express.json({ limit: '8kb' }));
 app.use(express.static(path.join(__dirname, 'public')));
 
 const androidInstallers = {
-  'Savewave-android-arm64.apk': 'https://github.com/kuberbassi/savewave/releases/latest/download/Savewave-android-arm64.apk',
-  'Savewave-android-armv7.apk': 'https://github.com/kuberbassi/savewave/releases/latest/download/Savewave-android-armv7.apk'
+  'Savewave-android-arm64.apk': 'https://github.com/kuberbassi/savewave/releases/latest/download/Savewave-android-arm64.apk'
 };
 app.get('/downloads/:filename', (req, res, next) => {
   const destination = androidInstallers[req.params.filename];